simple sample app cleanup + reverted deleted beep sound setup
