+++ Pdf417MobiSample/pdf417MobiSample/src/main/res/layout/activity_main.xml
@@ -8,13 +8,13 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".Pdf417MobiDemo" >
+    tools:context=".Pdf417MobiDemoActivity" >
 
     <Button
         android:id="@+id/btnScan"
         android:layout_width="142dp"
         android:layout_height="wrap_content"
-        android:onClick="btnScan_click"
+        android:onClick="onScanButtonClick"
         style="@android:style/Widget.Button"
         android:text="@string/StartScanning" />
